--- src/main/java/aws/iot/README.md
@@ -1,19 +1,0 @@
-# AWS IoT
-
-AWS IoT SDK for Java: http://docs.aws.amazon.com/iot/latest/developerguide/iot-sdks.html#iot-java-sdk
-
-AWS IoT SDK for Java on GitHub: https://github.com/aws/aws-iot-device-sdk-java
-
-### Install the SDK
-Install the SDK Using Maven
-
-```XML
-<dependencies>
-    <dependency>
-        <groupId>com.amazonaws</groupId>
-        <artifactId>aws-iot-device-sdk-java</artifactId>
-        <version>1.1.1</version>
-    </dependency>
-</dependencies>
-```
-
