+++ src/main/java/aws/iot/publisher.java
@@ -4,14 +4,10 @@ import com.amazonaws.services.iot.client.AWSIotException;
 import com.amazonaws.services.iot.client.AWSIotMqttClient;
 import com.amazonaws.services.iot.client.AWSIotQos;
 import com.amazonaws.services.iot.client.AWSIotTimeoutException;
-import com.amazonaws.services.iot.client.sample.sampleUtil.SampleUtil;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
 
 public class publisher {
   private static final AWSIotQos QOS = AWSIotQos.QOS0;
+  private static final long QUIESCE_TIMEOUT = 5000;
   private static final String TOPIC = "MyIoTButtonTopìc";
   private static final String PAYLOAD = "{\n" +
       "    \"deviceid\" : \"iot123\",\n" +
@@ -22,28 +18,13 @@ public class publisher {
       "        \"longitude\" : -122.3359976\n" +
       "    }\n" +
       "}";
-  private static final long QUIESCE_TIMEOUT = 5000;
 
   public static void main (String args[]) {
 
     if(args.length < 1) { showHelp(); }
 
-    Properties prop = new Properties();
-    ClassLoader loader = Thread.currentThread().getContextClassLoader();
-    InputStream input = loader.getResourceAsStream(args[0]);
+    AWSIotMqttClient client = new IoTMQTTClient(args[0]).getClient();
 
-    try {
-      prop.load(input);
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-
-    // SampleUtil.java and its dependency PrivateKeyReader.java can be copied from the sample source code.
-    // Alternatively, you could load key store directly from a file - see the example included in this README.
-    SampleUtil.KeyStorePasswordPair pair = SampleUtil.getKeyStorePasswordPair(prop.getProperty("certificateFile"), prop.getProperty( "privateKeyFile"));
-    AWSIotMqttClient client = new AWSIotMqttClient(prop.getProperty("clientEndpoint"), prop.getProperty("clientId"), pair.keyStore, pair.keyPassword);
-
-    // optional parameters can be set before connect()
     try {
       client.connect();
