+++ src/main/java/aws/iot/subscriber.java
@@ -1,13 +1,8 @@
 package aws.iot;
 
 import com.amazonaws.services.iot.client.*;
-import com.amazonaws.services.iot.client.sample.sampleUtil.SampleUtil;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
-
-public class subscriber extends AWSIotTopic {
+public class subscriber {
   private static final AWSIotQos QOS = AWSIotQos.QOS0;
   private static final String TOPIC = "MyIoTButtonTopìc";
 
@@ -15,40 +10,19 @@ public class subscriber extends AWSIotTopic {
 
     if(args.length < 1) { showHelp(); }
 
-    Properties prop = new Properties();
-    ClassLoader loader = Thread.currentThread().getContextClassLoader();
-    InputStream input = loader.getResourceAsStream(args[0]);
+    AWSIotMqttClient client = new IoTMQTTClient(args[0]).getClient();
 
-    try {
-      prop.load(input);
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-
-    // SampleUtil.java and its dependency PrivateKeyReader.java can be copied from the sample source code.
-    // Alternatively, you could load key store directly from a file - see the example included in this README.
-    SampleUtil.KeyStorePasswordPair pair = SampleUtil.getKeyStorePasswordPair(prop.getProperty("certificateFile"), prop.getProperty("privateKeyFile"));
-    AWSIotMqttClient client = new AWSIotMqttClient(prop.getProperty("clientEndpoint"), prop.getProperty("clientId"), pair.keyStore, pair.keyPassword);
-
-    // optional parameters can be set before connect()
     try {
       client.connect();
 
-      subscriber topic = new subscriber(TOPIC, QOS);
-      client.subscribe(topic);
+      MyTopic topic = new MyTopic(TOPIC, QOS);
+      client.subscribe(topic, 30000, true);
 
     } catch (AWSIotException e) {
       e.printStackTrace();
+    } catch (AWSIotTimeoutException e) {
+      e.printStackTrace();
     }
-  }
-
-  private subscriber(String topic, AWSIotQos qos) {
-    super(topic, qos);
-  }
-
-  @Override
-  public void onMessage(AWSIotMessage message) {
-    System.out.println(message.toString());
   }
 
   private static void showHelp()  {
